Moving all functions for scxml to jani to separate file
